Add options and conditionals to cater for CentOS Vagrant VM deployments

+++ vagrant/bootstrap_client.sh
@@ -4,18 +4,20 @@
 # Exit if the following file exists
 test -f /root/.provisioned && exit 0
 
-# Ensure we've got the latest APT caches
-apt-get update &> /dev/null
-
-# Ensure we have the required packages installed
-for i in puppetdb-terminus; do
-  dpkg -s ${i} &> /dev/null || apt-get -y install ${i} > /dev/null
-done
-
-# Ensure we have the required rubygems installed
-for i in hiera-eyaml; do
-  gem query --name ${i} --installed &> /dev/null || gem install --no-rdoc --no-ri ${i} > /dev/null
-done
+# Distribution-specific package considerations
+if [ -f /etc/redhat-release ]; then
+  yum makecache fast > /dev/null
+  yum install -y puppetdb-terminus > /dev/null
+  gem install --no-rdoc --no-ri hiera-eyaml > /dev/null
+else
+  apt-get update &> /dev/null
+  for i in puppetdb-terminus; do
+    dpkg -s ${i} &> /dev/null || apt-get -y install ${i} > /dev/null
+  done
+  for i in hiera-eyaml; do
+    gem query --name ${i} --installed &> /dev/null || gem install --no-rdoc --no-ri ${i} > /dev/null
+  done
+fi
 
 # Setup PuppetDB
 cat << EOF > /etc/puppet/puppetdb.conf
